added funtinoality to http for setting the mouse-control channels. also added a mean-calibrate routine. there seems to be some junk data in the testclient occasionally - I'm not sure why?  It always occurrs at the beginning of the trace, and occurrs on all channels, even the 'dead' channels 10 & 15; I'm not sure why.  perhaps it is ram corruption?  rx packets going there? or something even more sinister?

+++ trunk/firmware/usb.c
@@ -242,9 +242,9 @@ void do_IN3(void){
 	}
 	wreg(rEP3INBC,3);				// arm it
 }
-int calcMeanOfChannel(u32 u, u32 chan){
+int calcMeanOfChannel(u32 u, u8 chan){
 	u -= 64*32; 
-	u += chan * 2; 
+	u += (u32)(chan * 2); 
 	u &= 0x0003ffff ; // in case it wrapped. 
 	int sum = 0; 
 	int i;
